Improve lifecycle ordering
`componentDidMount` and `componentDidUpdate` are now called in correct order (`child1, child2, child3, parent`).

`getSnapshotBeforeUpdate` is still called in the wrong order (`parent, child1, child2, child3`). The fix for this
will not be too easy as it must be called before DOM manipulations and we do DOM manipulations while traversing down.
One possible approach would be to gather all DOM manipulations in an array and process them later similar as we do
for the DidMount/DidUpdate callbacks. The Q is whether that should already be done in `renderTree()` or as part of
`commit()`. When doing so in `renderTree()` we'd need to read DOM there for the child diffing, etc. and I don't think
this is a super good idea but I don't know.